Added full support for CRUD operations for school classes with respective consequences on operation - on edit change teachers if needed, on delete remove map to teacher and also delete all students in class

+++ Web/Gradebook.Web/Services/TeachersService.cs
@@ -99,6 +99,10 @@
             var teacher = _teachersRepository.All().FirstOrDefault(s => s.Id == id);
             if (teacher != null)
             {
+                if (teacher.Class != null)
+                {
+                    throw new ArgumentException($"Sorry, we cannot delete a teacher that is currently leading a class. Please first assign another lead teacher for the class.");
+                }
                 _teachersRepository.Delete(teacher);
                 await _teachersRepository.SaveChangesAsync();
             }
